# packages/ai-client/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://${string}` | (string & {});
};

/**
 * AssistantEvent
 *
 * One server-sent event of an answer.
 */
export type AssistantEvent = {
    /**
     * Event
     */
    event: ({
        type: 'text';
    } & TextEvent) | ({
        type: 'sources';
    } & SourcesEvent) | ({
        type: 'done';
    } & DoneEvent) | ({
        type: 'error';
    } & ErrorEvent);
};

/**
 * AssistantRequest
 */
export type AssistantRequest = {
    /**
     * Question
     */
    question: string;
};

/**
 * DocumentCreate
 */
export type DocumentCreate = {
    /**
     * Content
     *
     * Plain text.
     */
    content: string;
    /**
     * Title
     */
    title: string;
};

/**
 * DocumentOut
 */
export type DocumentOut = {
    /**
     * Chunkcount
     */
    chunkCount: number;
    /**
     * Createdat
     */
    createdAt: string;
    /**
     * Createdby
     */
    createdBy: string | null;
    /**
     * Error
     */
    error: string | null;
    /**
     * Id
     */
    id: string;
    /**
     * Status
     */
    status: 'pending' | 'indexing' | 'ready' | 'failed';
    /**
     * Title
     */
    title: string;
};

/**
 * DoneEvent
 */
export type DoneEvent = {
    /**
     * Type
     */
    type?: 'done';
    usage: Usage;
};

/**
 * ErrorEvent
 */
export type ErrorEvent = {
    /**
     * Code
     */
    code: 'AI_RUN_LIMIT' | 'UPSTREAM_UNAVAILABLE';
    /**
     * Type
     */
    type?: 'error';
};

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
    /**
     * Detail
     */
    detail?: Array<ValidationError>;
};

/**
 * HealthResponse
 */
export type HealthResponse = {
    /**
     * Status
     */
    status: 'ok';
};

/**
 * SentimentRequest
 */
export type SentimentRequest = {
    /**
     * Text
     *
     * Text to classify.
     */
    text: string;
};

/**
 * SentimentResponse
 */
export type SentimentResponse = {
    /**
     * Label
     */
    label: 'positive' | 'negative' | 'neutral';
    /**
     * Model
     *
     * Identifier of the model that produced the prediction.
     */
    model: string;
    /**
     * Score
     *
     * Confidence of the predicted label.
     */
    score: number;
};

/**
 * Source
 */
export type Source = {
    /**
     * Documentid
     */
    documentId: string;
    /**
     * Title
     */
    title: string;
};

/**
 * SourcesEvent
 */
export type SourcesEvent = {
    /**
     * Sources
     */
    sources: Array<Source>;
    /**
     * Type
     */
    type?: 'sources';
};

/**
 * TextEvent
 */
export type TextEvent = {
    /**
     * Text
     */
    text: string;
    /**
     * Type
     */
    type?: 'text';
};

/**
 * Usage
 */
export type Usage = {
    /**
     * Inputtokens
     */
    inputTokens: number;
    /**
     * Outputtokens
     */
    outputTokens: number;
};

/**
 * ValidationError
 */
export type ValidationError = {
    /**
     * Context
     */
    ctx?: {
        [key: string]: unknown;
    };
    /**
     * Input
     */
    input?: unknown;
    /**
     * Location
     */
    loc: Array<string | number>;
    /**
     * Message
     */
    msg: string;
    /**
     * Error Type
     */
    type: string;
};

export type LiveData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/health/live';
};

export type LiveResponses = {
    /**
     * Successful Response
     */
    200: HealthResponse;
};

export type LiveResponse = LiveResponses[keyof LiveResponses];

export type ReadyData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/health/ready';
};

export type ReadyResponses = {
    /**
     * Successful Response
     */
    200: HealthResponse;
};

export type ReadyResponse = ReadyResponses[keyof ReadyResponses];

export type AnswerData = {
    body: AssistantRequest;
    headers?: {
        /**
         * Authorization
         */
        authorization?: string | null;
        /**
         * X-Request-Id
         */
        'x-request-id'?: string | null;
    };
    path?: never;
    query?: never;
    url: '/v1/assistant/answers';
};

export type AnswerErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type AnswerError = AnswerErrors[keyof AnswerErrors];

export type AnswerResponses = {
    /**
     * Server-sent events, each `data:` line one AssistantEvent.
     */
    200: AssistantEvent;
};

export type AnswerResponse = AnswerResponses[keyof AnswerResponses];

export type ListDocumentsData = {
    body?: never;
    headers?: {
        /**
         * Authorization
         */
        authorization?: string | null;
        /**
         * X-Request-Id
         */
        'x-request-id'?: string | null;
    };
    path?: never;
    query?: never;
    url: '/v1/documents';
};

export type ListDocumentsErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type ListDocumentsError = ListDocumentsErrors[keyof ListDocumentsErrors];

export type ListDocumentsResponses = {
    /**
     * Response Listdocuments
     *
     * Successful Response
     */
    200: Array<DocumentOut>;
};

export type ListDocumentsResponse = ListDocumentsResponses[keyof ListDocumentsResponses];

export type CreateDocumentData = {
    body: DocumentCreate;
    headers?: {
        /**
         * Authorization
         */
        authorization?: string | null;
        /**
         * X-Request-Id
         */
        'x-request-id'?: string | null;
    };
    path?: never;
    query?: never;
    url: '/v1/documents';
};

export type CreateDocumentErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type CreateDocumentError = CreateDocumentErrors[keyof CreateDocumentErrors];

export type CreateDocumentResponses = {
    /**
     * Successful Response
     */
    201: DocumentOut;
};

export type CreateDocumentResponse = CreateDocumentResponses[keyof CreateDocumentResponses];

export type DeleteDocumentData = {
    body?: never;
    headers?: {
        /**
         * Authorization
         */
        authorization?: string | null;
        /**
         * X-Request-Id
         */
        'x-request-id'?: string | null;
    };
    path: {
        /**
         * Document Id
         */
        document_id: string;
    };
    query?: never;
    url: '/v1/documents/{document_id}';
};

export type DeleteDocumentErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type DeleteDocumentError = DeleteDocumentErrors[keyof DeleteDocumentErrors];

export type DeleteDocumentResponses = {
    /**
     * Successful Response
     */
    204: void;
};

export type DeleteDocumentResponse = DeleteDocumentResponses[keyof DeleteDocumentResponses];

export type SentimentData = {
    body: SentimentRequest;
    headers?: {
        /**
         * Authorization
         */
        authorization?: string | null;
        /**
         * X-Request-Id
         */
        'x-request-id'?: string | null;
    };
    path?: never;
    query?: never;
    url: '/v1/sentiment';
};

export type SentimentErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type SentimentError = SentimentErrors[keyof SentimentErrors];

export type SentimentResponses = {
    /**
     * Successful Response
     */
    200: SentimentResponse;
};

export type SentimentResponse2 = SentimentResponses[keyof SentimentResponses];